+++ autd3-link-soem/src/lib.rs
@@ -1,7 +1,7 @@
 #[cfg(feature = "local")]
 pub mod local;
 #[cfg(feature = "local")]
-pub use local::{EthernetAdapters, SyncMode, SOEM};
+pub use local::{EthernetAdapters, Status, SyncMode, SOEM};
 
 #[cfg(feature = "remote")]
 pub mod remote;