+++ autd3-link-soem/src/local/mod.rs
@@ -1,11 +1,14 @@
+mod builder;
 mod error;
 mod error_handler;
 mod ethernet_adapters;
 mod iomap;
 pub mod link_soem;
+mod sleep;
 mod soem_bindings;
 mod state;
 
 pub use autd3_driver::sync_mode::SyncMode;
+pub use error_handler::Status;
 pub use ethernet_adapters::EthernetAdapters;
 pub use link_soem::SOEM;