+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 21.0.2
 
 - Fix [#7](https://github.com/shinolab/autd3-rs/issues/7): `on_lost` and `on_err` callback don't work
+- Add `with_err_handler` to `autd3-link-soem::SOEM`
 
 # 21.0.1